refactor<CoreDependencyInjection>: Interface 제거 후 implement로 이동
