+++ Projects/Core/DependencyInjection/Sources/DIContainer/ResolveError.swift
@@ -1,6 +1,6 @@
 //
 //  ResolveError.swift
-//  CoreDependencyInjectionInterface
+//  CoreDependencyInjection
 //
 //  Created by Derrick kim on 4/18/24.
 //
