+++ Projects/Core/DependencyInjection/Sources/Injected.swift
@@ -1,6 +1,6 @@
 //
 //  Injected.swift
-//  Feature
+//  CoreDependencyInjection
 //
 //  Created by Derrick kim on 4/17/24.
 //
