+++ Projects/Core/DependencyInjection/Sources/SafeInjected.swift
@@ -1,6 +1,6 @@
 //
 //  SafeInjected.swift
-//  Feature
+//  CoreDependencyInjection
 //
 //  Created by Derrick kim on 4/17/24.
 //
